:recycle: refactor buttons-container
