+++ src/components/Form/ButtonsContainer.tsx
@@ -1,0 +1,11 @@
+import { Row } from 'antd'
+import styled from 'styled-components'
+
+const ButtonsContainer = styled(Row)`
+  width: 100%;
+  margin: auto;
+  margin-top: 50px;
+  margin-bottom: 44px;
+`
+
+export default ButtonsContainer
